requirements.txt
Django==5.0.4
pillow==10.3.0
django-environ==0.11.2
django-ckeditor-5==0.2.12
pytils==0.4.1
django-recaptcha==4.0.0
django-taggit==5.0.1
django-mptt==0.16.0